third_party/blink/web_tests/fast/js/navigator-language-expected.txt

Check return value of navigator.language.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS language is 'en-us'
PASS successfullyParsed is true

TEST COMPLETE
